% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{AssignmentMarks}
\alias{AssignmentMarks}
\title{Get a student's marks on assignments.}
\usage{
AssignmentMarks(ID = 111111111, date = Sys.Date())
}
\arguments{
\item{ID}{A student's ID number (should be a 9 digit integer).}

\item{date}{A \code{\link{date}} class object.}
}
\value{
A matrix with 3 columns:
\enumerate{
   \item The student's marks on assignments.
   \item The full potential marks.
   \item The student's marks as fractions of the full marks.
}
}
\description{
Enter a student's ID and date.  Function uses \code{\link{IDToAssignmentMark}}.
Returns a matrix of a student's marks on assignments up to that date.
}
\seealso{
\code{\link{date}}

\code{\link{IDToAssignmentMark}}

\code{\link{TestMarks}}
}

