% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{basicOutputs}
\alias{basicOutputs}
\alias{listTables}
\alias{readAllInfo}
\alias{readAssignments}
\alias{readClassParticipation}
\alias{readLongformGrades}
\alias{readMCAnswers}
\alias{readStudents}
\alias{showTable}
\title{Basic outputs}
\usage{
listTables(conn = DBconn())

showTable(table, conn = DBconn())

readStudents(conn = DBconn())

readAssignments(conn = DBconn())

readMCAnswers(conn = DBconn())

readLongformGrades(conn = DBconn())

readClassParticipation(conn = DBconn())

readAllInfo(conn = DBconn())
}
\arguments{
\item{conn}{An SQL connection to a database file.}

\item{table}{The name of a table to be shown.  Enter in quotations ("table").}
}
\description{
List course.db's tables and return their contents.
\describe{
   \item{listTables()}{Returns a vector of strings containing the names of
         the tables for reference.}
   \item{showTable()}{Takes one of the strings returned by listTables() as
         an input and returns that table's contents.}
   \item{readStudents()}{Returns the contents of the "students" table.}
   \item{readAssignments()}{Returns the contents of the "assignments" table.}
   \item{readMCAnswers()}{Returns the contents of the "mcAnswers" table.}
   \item{readLongFormGrades()}{Returns the contents of the "longFormGrades" table.}
   \item{classParticipation()}{Returns the contents of the "classParticipation" table.}
   \item{readAllInfo()}{Returns the contents of all tables.}
}
}
\seealso{
\code{\link{createDB}}

\code{\link{DBconn}}
}

