% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createDB.R
\name{createDB}
\alias{createDB}
\title{Create initial database.}
\usage{
createDB(conn = DBconn())
}
\arguments{
\item{conn}{An SQL connection to a database file.}
}
\value{
createDB does not return a value.
}
\description{
Create initial database; only run once.
Afterwards, use other functions to update the database.
}
\examples{
createDB()
createDB(DBconn())
createDB(connectionToDatabaseName)
}
\seealso{
\code{\link{DBconn}}
}

