% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{IDAndExamNumberToGrade}
\alias{IDAndExamNumberToGrade}
\title{Get a given student's grade on a given exam.}
\usage{
IDAndExamNumberToGrade(ID = 111111111, examNumber = "2")
}
\arguments{
\item{ID}{A student's ID number (should be a 9 digit integer).}

\item{examNumber}{A string, the number (or name) assigned to an exam.}
}
\value{
A vector with 3 elements.
\enumerate{
   \item The student's score.
   \item The full potential score.
   \item The student's score as a fraction of the full score.
}
}
\description{
Enter a student's ID and the name (or 'examNumber') of an exam.
Returns their score, the score on a perfectly completed exam, and their
score as a fraction.
}
\seealso{
\code{\link{IDToAssignmentMark}}
}

