% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{IDToAssignmentMark}
\alias{IDToAssignmentMark}
\title{Get a given student's mark on a given assignment.}
\usage{
IDToAssignmentMark(ID = 111111111, assignmentNumber = "1")
}
\arguments{
\item{ID}{A student's ID number (should be a 9 digit integer).}

\item{assignmentNumber}{A string, the number (or name) of an assignment.}
}
\value{
A vector with 3 elements.
\enumerate{
   \item The student's score.
   \item The full potential score.
   \item The student's score as a fraction of the full score.
}
}
\description{
Enter a student's ID and the name (or 'examNumber') of an assignment.
Returns their score, the score on a perfectly completed assignment, and their
score as a fraction.
}
\seealso{
\code{\link{IDAndExamNumberToGrade}}
}

