% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{IDToAttendance}
\alias{IDToAttendance}
\title{Get a given student's attendance grade.}
\usage{
IDToAttendance(ID = 111111111, date = Sys.Date(),
  attendanceMethod = "toDate")
}
\arguments{
\item{ID}{A student's ID number (should be a 9 digit integer).}

\item{date}{A \code{\link{date}} class object.}

\item{attendanceMethod}{One of
\enumerate{
      \item A string, "toDate" -- evaluate attendance as a fraction of
            all dates attended by the student out of all unique dates
            entered in the "classParticipation" table up until the date
            parameter.
      \item A numeric value --  evaluate attendance by dividing the
            number of dates attended up until the date parameter by the
            numeric value entered, capped at 1.
}}
}
\value{
A numeric value.
}
\description{
Enter a student's ID, the date up to which to count attendance,
and an evaluation method.  Returns their attendance grade as evaluated by
the chosen method.
}
\seealso{
\code{\link{date}}

\code{\link{IDToClassParticipation}}
}

