% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{IDToClassParticipation}
\alias{IDToClassParticipation}
\title{Get a class participation grade by weighting the attendance mark and the
questions asked and/ or answered mark.}
\usage{
IDToClassParticipation(ID = 111111111, date = Sys.Date(),
  cpWeighting = c(0.5, 0.5), attendanceMethod = "toDate",
  questionMethod = c("answer", "percent"))
}
\arguments{
\item{ID}{A student's ID number (should be a 9 digit integer).}

\item{date}{A \code{\link{date}} class object.}

\item{cpWeighting}{A 2-element vector: \enumerate{
\item The weight to assign to attendance.
\item the weight to assign to questions.
} The mark returned will be incorrect if the cpWeighting vector does not
sum to 1.}

\item{attendanceMethod}{Either a string, "toDate", or a numeric value;
see \code{\link{IDToAttendance}}.}

\item{questionMethod}{A 2-element vector; \enumerate{
\item A string -- either "ask", "answer", or "both", depending on
      which function you want to use to evaluate questions for the
      class participation grade.
\item Either a string, "percent", or a numeric value.  See the Arguments
      in \code{\link{IDToQuestionsAsked}}, for an explanation.
}}
}
\value{
A numeric value, the fraction of a full grade the student receives.
}
\description{
Enter a student's ID, the date up to which to count, a weighting of
attendance vs. participation, a method for evaluating attendance, and a
method for evaluating questions.  The function uses
\code{\link{IDToAttendance}}, and one of
\code{\link{IDToQuestionsAnswered}},  \code{\link{IDToQuestionsAsked}}, and
\code{\link{IDToCombinedQuestions}}.  Returns a mark according to the
parameters chosen.
}
\seealso{
\code{\link{date}}

\code{\link{IDToAttendance}}

\code{\link{IDToQuestionsAsked}}

\code{\link{IDToQuestionsAnswered}}

\code{\link{IDToCombinedQuestions}}
}

