% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{IDToCurrentGrade}
\alias{IDToCurrentGrade}
\title{Get student's overall grade.}
\usage{
IDToCurrentGrade(ID = 111111111, totalWeighting = c(0.2, 0, 0.3, 0.5),
  date = Sys.Date(), cpWeighting = c(0.5, 0.5),
  attendanceMethod = "toDate", questionMethod = c("answer", "percent"))
}
\arguments{
\item{ID}{A student's ID number (should be a 9 digit integer).}

\item{totalWeighting}{A 4-element vector: \enumerate{
\item The weight to assign to assignments.
\item the weight to assign to class participation.
\item The weight to assign to all tests other than the last one.
\item The weight to assign to the final exam.
} The overall grade will be off if the totalWeighting vector does not
sum to 1.}

\item{date}{A \code{\link{date}} class object.}

\item{cpWeighting}{A 2-element vector: \enumerate{
\item The weight to assign to attendance.
\item the weight to assign to questions.
} The mark returned will be incorrect if the cpWeighting vector does not
sum to 1.}

\item{attendanceMethod}{Either a string, "toDate", or a numeric value;
see \code{\link{IDToAttendance}}.}

\item{questionMethod}{A 2-element vector; \enumerate{
\item A string -- either "ask", "answer", or "both", depending on
      which function you want to use to evaluate questions for the
      class participation grade.
\item Either a string, "percent", or a numeric value.  See the Arguments
      in \code{\link{IDToQuestionsAsked}}, for an explanation.
}}
}
\value{
A numeric value, the fraction of a full grade the student receives.
}
\description{
Calculates and returns a student's overall grade using functions
\code{\link{IDToClassParticipation}}, \code{\link{AssignmentMarks}},
and \code{\link{TestMarks}}, their inputs, and a vector of weights to
assign to assignments, class participation, (other) tests, and the final exam.
The number will be off if the totalWeighting vector does not add up to 1.
}
\seealso{
\code{\link{date}}

\code{\link{IDToClassParticipation}}

\code{\link{AssignmentMarks}}

\code{\link{TestMarks}}
}

