% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{IDToNotes}
\alias{IDToNotes}
\title{Retrieve all notes on a given student.}
\usage{
IDToNotes(ID = 111111111)
}
\arguments{
\item{ID}{A student's ID number (should be a 9 digit integer).}
}
\value{
A data.frame with a single column.
}
\description{
Enter the student's ID.  Returns a data.frame with any notes entered.
}
\seealso{
\code{\link{readStudents}}
}

