% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{IDToQuestionsAsked}
\alias{IDToQuestionsAsked}
\title{Get a mark based off of the number of questions a student has asked.}
\usage{
IDToQuestionsAsked(ID = 111111111, date = Sys.Date(),
  evalMethod = "percent")
}
\arguments{
\item{ID}{A student's ID number (should be a 9 digit integer).}

\item{date}{A \code{\link{date}} class object.}

\item{evalMethod}{One of
\enumerate{
      \item A string, "percent" -- determine the student's grade as a
            fraction with the number of questions the student has asked
            as the numerator and the of the maximum number of questions
            asked by a single student in the table as the denominator.
      \item A numeric value --  if the student has asked at least that
            number of questions, they get full marks; otherwise, none.
}}
}
\value{
A vector with 2 or 3 elements.
\enumerate{
   \item The number of questions the student asked.
   \item The student's mark as a fraction.
   \item If a threshold number was given as the evalMethod, the threshold
         for full marks.
}
}
\description{
Enter a student's ID, the date up to which to count, and an evaluation
method.  Returns a mark according to the evaluation method chosen.
}
\seealso{
\code{\link{date}}

\code{\link{IDToClassParticipation}}

\code{\link{IDToQuestionsAnswered}}

\code{\link{IDToCombinedQuestions}}
}

