% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{NameToID}
\alias{NameToID}
\title{Get student's ID number from their name.}
\usage{
NameToID(givenNames = "deedoublename", lastName = "cee")
}
\arguments{
\item{givenNames}{A string, a student's first name (and any middle names used).}

\item{lastName}{A string,  that student's last name.}
}
\value{
A dataframe containing the IDs, given names, and last names of all matches.
}
\description{
Enter a student's name -- given names in the first string, last name in the second.
NameToID() will return a dataframe with the IDs and names of all students
with names similar (according to \link{agrep()}) to the strings the user
gives.
}

