% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entries.R
\name{NewAssignmentEntry}
\alias{NewAssignmentEntry}
\title{Enter a new row into the "assignments" table.}
\usage{
NewAssignmentEntry(ID = 999999999, assignmentNumber, date = Sys.Date(),
  grade)
}
\arguments{
\item{ID}{A student's ID number (should be a 9 digit integer).}

\item{assignmentNumber}{A string, the number (or name) of an assignment.}

\item{date}{A \code{\link{date}} class object.}

\item{grade}{A numeric value, the grade the student achieved on the assignment.}
}
\description{
Enter a new row into the "assignments" table.
}
\seealso{
\code{\link{date}}

Other data.entry.functions: \code{\link{NewCPEntry}};
  \code{\link{NewLFEntry}}; \code{\link{NewMCEntry}};
  \code{\link{NewStudentEntry}}
}

