% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entries.R
\name{NewCPEntry}
\alias{NewCPEntry}
\title{Enter a new row into the "classParticipation" table.}
\usage{
NewCPEntry(ID = 999999999, date = Sys.Date(), attended = TRUE,
  questionAnswered = "", questionAsked = "", participationNotes = "")
}
\arguments{
\item{ID}{A student's ID number (should be a 9 digit integer).}

\item{date}{A \code{\link{date}} class object.}

\item{attended}{A boolean value, whether the student attended on that date.}

\item{questionAnswered}{A string, the question the student answered.}

\item{questionAsked}{A string, the question the student asked.}

\item{notes}{A string, any notes you wish to include.}
}
\description{
Enter a new row into the "classParticipation" table.
}
\seealso{
\code{\link{date}}

Other data.entry.functions: \code{\link{NewAssignmentEntry}};
  \code{\link{NewLFEntry}}; \code{\link{NewMCEntry}};
  \code{\link{NewStudentEntry}}
}

