% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entries.R
\name{NewLFEntry}
\alias{NewLFEntry}
\title{Enter a new row into the "longFormGrades" table.}
\usage{
NewLFEntry(ID = 999999999, grade, questionNumber, examNumber, examCode,
  date = Sys.Date())
}
\arguments{
\item{ID}{A student's ID number (should be a 9 digit integer).}

\item{grade}{A numeric value, the grade on a question achieved by the student.}

\item{questionNumber}{The number of the question on the test.}

\item{examNumber}{A string, the number (or name) assigned to an exam.}

\item{examCode}{A string, typically a 3-digit integer, but exam versions
can be distinguished from each other using names.}
}
\description{
Enter a new row into the "longFormGrades" table.
}
\seealso{
Other data.entry.functions: \code{\link{NewAssignmentEntry}};
  \code{\link{NewCPEntry}}; \code{\link{NewMCEntry}};
  \code{\link{NewStudentEntry}}
}

