% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entries.R
\name{NewMCEntry}
\alias{NewMCEntry}
\title{Enter a new row into the "mcAnswers" table.}
\usage{
NewMCEntry(ID = 999999999, answer, questionNumber, questionValue = 2,
  examNumber, examCode, date = Sys.Date())
}
\arguments{
\item{ID}{A student's ID number (should be a 9 digit integer).}

\item{answer}{An integer, the multiple choice answer given by the student.}

\item{questionNumber}{The number of the question on the test.}

\item{questionValue}{A numeric value, the marks that question is worth.}

\item{examNumber}{A string, the number (or name) assigned to an exam.}

\item{examCode}{A string, typically a 3-digit integer, but exam versions
can be distinguished from each other using names.}
}
\description{
Enter a new row into the "mcAnswers" table.
}
\seealso{
Other data.entry.functions: \code{\link{NewAssignmentEntry}};
  \code{\link{NewCPEntry}}; \code{\link{NewLFEntry}};
  \code{\link{NewStudentEntry}}
}

