% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entries.R
\name{NewStudentEntry}
\alias{NewStudentEntry}
\title{Enter a new row into the "students" table.}
\usage{
NewStudentEntry(ID, email, lastName, givenNames, program, notes)
}
\arguments{
\item{ID}{A student's ID number (should be a 9 digit integer).}

\item{email}{A string, an email address.}

\item{lastName}{A string, the student's last name.}

\item{givenNames}{A string, that student's first name (and any middle names used).}

\item{program}{A string, a program name.  Only use a few standardized names.  You don't want "econ", "Econ", "economics", etc.}

\item{notes}{A string, any notes you wish to include.}
}
\description{
Enter a new row into the "students" table.
}
\seealso{
Other data.entry.functions: \code{\link{NewAssignmentEntry}};
  \code{\link{NewCPEntry}}; \code{\link{NewLFEntry}};
  \code{\link{NewMCEntry}}
}

